[noci] fixed git cmd
