+++ tool/cmds/lint.py
@@ -1,14 +1,14 @@
 from typing import Iterable
 
 from cmd import CmdContext
+from cmds.shared import print_stdout
 from dart import DartFormatConf
 
 
 def lint(ctx: CmdContext):
     conf = ctx.proj.settings.get("cmd.lint.conf", DartFormatConf)
-    proc = ctx.proj.runner.format(conf)
-    for line in proc.stdout:
-        ctx.term << line.decode("utf-8").splitlines()[0]
+    proc = ctx.proj.dartRunner.format(conf)
+    print_stdout(ctx, proc)
 
 
 class LintCmd:
