+++ tool/dart.py
@@ -1,9 +1,10 @@
-import shlex
-from typing import Union, Optional, Sequence
+from typing import Union, Optional
 
-from filesystem import File, Pathable, Directory
+from filesystem import File, Pathable
 
 import subprocess
+
+from runner import Runner
 
 
 class DartFi(File):
@@ -39,23 +40,8 @@ class DartFormatConf:
 
 
 class DartRunner:
-    def __init__(self, root: Directory):
-        self.root = root
-
-    def run(self, *, full: str = None, seq: Sequence[str] = None) -> subprocess.Popen:
-        if full is not None:
-            args = shlex.split(full)
-        elif seq is not None:
-            args = seq
-        else:
-            raise Exception("no full or seq given as command line args")
-        return subprocess.Popen(
-            args=args,
-            bufsize=-1, shell=True,
-            cwd=self.root.abs_path,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE
-        )
+    def __init__(self, runner: Runner):
+        self.runner = runner
 
     def format(self, config: DartFormatConf) -> subprocess.Popen:
         args = ["dart", "format", "."]
@@ -63,4 +49,4 @@ class DartRunner:
         if length is not None:
             args.append("-l")
             args.append(str(length))
-        return self.run(seq=args)
+        return self.runner.run(seq=args)
